move common Program to core
